moved get_or_create to util, updated cell init
