+++ backend/api/utils/import_cell_data.py
@@ -21,7 +21,7 @@ from sqlalchemy.orm import Session
 from ..conn import engine
 from ..models.power_data import PowerData
 from ..models.teros_data import TEROSData
-from ..get_or_create import get_or_create_cell, get_or_create_logger
+from .get_or_create import get_or_create_cell, get_or_create_logger
 
 
 def import_cell_data(path, logger_name, cell_name, batch_size=10000):
